refactor: move flag parsers to shared cli-helpers
parseSize and parsePositiveInt are generic CLI flag parsers, not
migrate concepts; they move next to formatFileSize (parseSize's
inverse) in src/utils/cli-helpers.ts so the next command with a size
or count flag reuses them instead of writing a third parser.
parseCidList stays in migrate's util: it parses migrate's input file
format. Folding payments/setup.ts's TiB-per-month regex into
parseSize is a possible follow-up.
